Feat(jobOffer): adding security a job offer can only be deleted,created or updated by the authenticated owner.

+++ service/src/main/java/be/portal/job/services/impls/JobOfferServiceImpl.java
@@ -2,17 +2,19 @@ package be.portal.job.services.impls;
 
 import be.portal.job.dtos.jobOffer.requests.JobOfferPostRequest;
 import be.portal.job.dtos.jobOffer.responses.JobOfferResponse;
+import be.portal.job.entities.JobAdvertiser;
 import be.portal.job.entities.JobOffer;
 import be.portal.job.repositories.CompanyAdvertiserRepository;
 import be.portal.job.repositories.ContractTypeRepository;
 import be.portal.job.repositories.JobFunctionRepository;
 import be.portal.job.repositories.JobOfferRepository;
-import be.portal.job.services.CompanyAdvertiserService;
 import be.portal.job.services.JobOfferService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -22,7 +24,6 @@ public class JobOfferServiceImpl implements JobOfferService {
     private final JobFunctionRepository jobFunctionRepository;
     private final ContractTypeRepository contractTypeRepository;
     private final CompanyAdvertiserRepository companyAdvertiserRepository;
-
 
     @Override
     public List<JobOfferResponse> getAll() {
@@ -42,6 +43,13 @@ public class JobOfferServiceImpl implements JobOfferService {
     public JobOfferResponse deleteJobOffer(Long id) {
         return jobOfferRepository.findById(id)
                 .map(jobOffer -> {
+                    JobAdvertiser currentUser = (JobAdvertiser) SecurityContextHolder.getContext()
+                            .getAuthentication().getPrincipal();
+
+                    if (!Objects.equals(currentUser.getId(), jobOffer.getAgent().getJobAdvertiser().getId())
+                            && currentUser.getAuthorities().contains("ADMIN")) {
+                        throw new RuntimeException("You are not allowed to delete job offers for other job advertisers");
+                    }
                     jobOfferRepository.delete(jobOffer);
                     return JobOfferResponse.fromEntity(jobOffer);
                 })
@@ -53,6 +61,13 @@ public class JobOfferServiceImpl implements JobOfferService {
         JobOffer jobOffer = new JobOffer();
         jobOfferPostRequest.toEntity(jobOffer);
         jobOffer.setAgent(companyAdvertiserRepository.findById(jobOfferPostRequest.agentId()).orElseThrow());
+        JobAdvertiser currentUser = (JobAdvertiser) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+
+        if (!Objects.equals(currentUser.getId(), jobOffer.getAgent().getJobAdvertiser().getId())
+                && currentUser.getAuthorities().contains("ADMIN")) {
+            throw new RuntimeException("You are not allowed to create job offers for other job advertisers");
+        }
         jobOffer.setJobFunction(jobFunctionRepository.findById(jobOfferPostRequest.jobFunctionId()).orElseThrow());
         jobOffer.setContractType(contractTypeRepository.findById(jobOfferPostRequest.contractTypeId()).orElseThrow());
         return JobOfferResponse.fromEntity(jobOfferRepository.save(jobOffer));
@@ -62,6 +77,13 @@ public class JobOfferServiceImpl implements JobOfferService {
     public JobOfferResponse updateJobOffer(Long id, JobOfferPostRequest jobOfferPostRequest) {
         JobOffer jobOffer = jobOfferRepository.findById(id).orElseThrow();
         jobOfferPostRequest.toEntity(jobOffer);
+        JobAdvertiser currentUser = (JobAdvertiser) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+
+        if (!Objects.equals(currentUser.getId(), jobOffer.getAgent().getJobAdvertiser().getId())
+                && currentUser.getAuthorities().contains("ADMIN")) {
+            throw new RuntimeException("You are not allowed to update job offers for other job advertisers");
+        }
         jobOffer.setAgent(companyAdvertiserRepository.findById(jobOfferPostRequest.agentId()).orElseThrow());
         jobOffer.setJobFunction(jobFunctionRepository.findById(jobOfferPostRequest.jobFunctionId()).orElseThrow());
         jobOffer.setContractType(contractTypeRepository.findById(jobOfferPostRequest.contractTypeId()).orElseThrow());
